+++ client/src/shared/constants/routes.ts
@@ -1,0 +1,10 @@
+export const navbarRoutes = {
+  home: {
+    title: 'Головна',
+    path: '/',
+  },
+};
+
+export const otherRoutes = {};
+
+export const routes = Object.assign(navbarRoutes, otherRoutes);