+++ client/src/theme/override.ts
@@ -1,8 +1,9 @@
 import type { ThemeConfig } from 'antd';
-import { Raleway, Source_Serif_4 } from 'next/font/google';
+import { Raleway, Source_Sans_3 } from 'next/font/google';
+import { colorBorder } from '.';
 
 const raleway = Raleway({ subsets: ['latin', 'cyrillic-ext'] });
-const sourceSerif4 = Source_Serif_4({ subsets: ['latin', 'cyrillic-ext'] });
+const sourceSans3 = Source_Sans_3({ subsets: ['latin', 'cyrillic-ext'] });
 
 export const theme: ThemeConfig = {
   token: {
@@ -11,7 +12,7 @@ export const theme: ThemeConfig = {
     colorTextSecondary: '#646464',
     colorText: '#3c3b3b',
     fontFamily: raleway.style.fontFamily,
-    colorBorder: 'rgba(160, 160, 160, 0.3)',
+    colorBorder,
   },
   components: {
     Button: {
@@ -22,7 +23,13 @@ export const theme: ThemeConfig = {
       fontSizeHeading2: 20,
     },
     Card: {
-      fontFamily: sourceSerif4.style.fontFamily,
+      fontFamily: sourceSans3.style.fontFamily,
+    },
+    Layout: {
+      colorBgHeader: '#fff',
+    },
+    Input: {
+      fontFamily: sourceSans3.style.fontFamily,
     },
   },
 };