+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 Всі важливі зміни в цьому проєкті будуть задокументовані в цьому файлі.
 
+## [0.0.2] - 2023-08-02
+
+### Added
+
+#### Client
+
+- Mobile version of navbar at the top of the page
+- Navbar drawer to hide links on the mobile
+
+#### Backend
+
+- Main entities of the app
+- Data to have the content to fetch on the client
+
 ## [0.0.1] - 2023-07-31
 
 ### Added